backend/requirements.txt
langchain
langchain-google-genai
python-dotenv
langchain-community
langchain-ollama
langgraph
langchain-core
langchainhub
FastAPI
uvicorn[standard]
pydantic
python-jose
bcrypt==3.2.2
passlib==1.7.4
python-multipart
sqlalchemy
cryptography
pyodbc
sqlglot